+++ backend/src/model/activity.model.ts
@@ -7,23 +7,23 @@ import pool from '../config/db';
 import ErrorDetail from './error.model';
 import "dotenv/config"; 
 
-interface ActivityRefType{
+export interface ActivityRefType{
   id: number,
   name: string, 
-  description: string | null,
+  description: {[key: string]: any} | null,
   cluster: number
 }
 
-interface ActivityJunctionType{
+export interface ActivityJunctionType{
   id: number,
   activity_id: number,
   user_id: string,
   timestamp: string,
   ip: string,
-  detail: string | null
+  detail: {[key: string]: any} | null
 }
 
-interface ActivityJoinType{
+export interface ActivityJoinType{
   id: number,
   activity_id: number,
   name: string,
@@ -31,7 +31,21 @@ interface ActivityJoinType{
   user_id: string,
   timestamp: string,
   ip: string,
-  detail: string | null
+  detail: {[key: string]: any} | null
+}
+
+export interface ActivityJoinUserType{
+  id: number,
+  activity_id: number,
+  name: string,
+  description: string | null,
+  user_id: string,
+  timestamp: string,
+  ip: string,
+  detail: {[key: string]: any} | null,
+  user_name: string, 
+  nip: string, 
+  kppn: string
 }
 
 class Activity{
@@ -98,10 +112,29 @@ class Activity{
     }
   }
 
-  async createActivity(userId: string, type: number, ip: string, detail: string|number|null = null){
+  async getByTypeDates(type: number[], startDate: Date, endDate: Date): Promise<ActivityJoinUserType[]>{ //Use date.toISOString() function , ex: 2023-01-01 00:00:00.
     try{
-      const q = `INSERT INTO activity_junction (activity_id, user_id, ip, detail) VALUES ($1, $2, $3, $4) RETURNING *`;
-      const result = await pool.query(q, [type, userId, ip, detail]);
+      const q = `SELECT activity_junction.*, activity_ref.name, activity_ref.description, user_ref.name AS user_name, user_ref.username AS nip, user_ref.kppn
+                 FROM activity_junction
+                 INNER JOIN activity_ref
+                 ON activity_junction.activity_id = activity_ref.id
+                 LEFT JOIN user_ref
+                 ON user_ref.username = activity_junction.user_id 
+                 WHERE activity_junction.activity_id = ANY($1)
+                 AND activity_junction.timestamp BETWEEN $2 AND $3
+                 ORDER BY id ASC`;
+      const result = await pool.query(q, [type, startDate, endDate]);
+      return result.rows
+    }catch(err){
+      throw err
+    }
+  }
+
+  async createActivity(userId: string, type: number, ip: string, detail: {}|string|number|null = null){
+    try{
+      const parsedDetail = detail !== null ? JSON.stringify(detail) : null;
+      const q = `INSERT INTO activity_junction (activity_id, user_id, ip, detail) VALUES ($1, $2, $3, $4::jsonb) RETURNING *`;
+      const result = await pool.query(q, [type, userId, ip, parsedDetail]);
       return result.rows
     }catch(err){
       throw err